requirements.txt
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==5.0.1
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.8
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
commonmark==0.9.1
contourpy==1.0.5
cycler==0.10.0
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0
django-ckeditor==6.5.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.7.0
django-js-asset==2.0.0
django-leaflet==0.27.1
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
drf-spectacular==0.21.0
et-xmlfile==1.1.0
Fiona @ file:///C:/Users/cfick/pipwin/Fiona-1.8.20-cp38-cp38-win_amd64.whl
Flask==1.1.2
folium==0.12.1.post1
fonttools==4.28.5
frozenlist==1.2.0
future==0.18.2
GDAL @ file:///C:/Users/cfick/pipwin/GDAL-3.3.3-cp38-cp38-win_amd64.whl
geocoder==1.38.1
geographiclib==1.52
geoip2==4.5.0
geopandas==0.11.1
geopy==2.2.0
geos==0.2.2
gunicorn==20.1.0
h11==0.12.0
hjson==3.1.0
html5lib==1.1
httpcore==0.15.0
httpx==0.23.0
humanfriendly==10.0
idna==2.10
importlib-metadata==5.0.0
importlib-resources==5.4.0
inflection==0.5.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
Js2Py==0.71
jsonfield==3.1.0
jsonschema==4.2.1
keyring==23.9.3
kiwisolver==1.4.4
lxml==4.6.2
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.1
maxminddb==2.2.0
more-itertools==8.14.0
multidict==5.2.0
munch==2.5.0
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pipwin==0.5.1
pkginfo==1.8.3
psutil==5.9.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycountry==22.3.5
Pygments==2.13.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
PyPrind==2.11.3
pyproj==3.4.0
pyreadline3==3.4.1
pyrsistent==0.18.0
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
ratelim==0.1.6
readme-renderer==37.2
reportlab==3.5.66
requests==2.27.0
requests-toolbelt==0.10.0
rfc3986==1.5.0
rich==12.6.0
Rtree==0.9.7
Shapely==1.7.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.3.1
starlette==0.21.0
tablib==3.1.0
twine==4.0.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.3
uuid==1.30
uvicorn==0.18.3
webencodings==0.5.1
websockets==10.3
Werkzeug==1.0.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.6.0
